Add StoreMember instruction to bytecode, to improve code generation

+++ cobrac/src/interpreter/tests.rs
@@ -8,7 +8,7 @@ use super::value::Value;
 pub struct Test
 {
     pub name: String,
-    pub ret: i64,
+    pub ret: isize,
     pub code: String,
 }
 
@@ -24,12 +24,12 @@ impl Test
         let ret: String = data.chars().skip(5).take_while(|c| c.is_numeric()).collect();
         Test{
             name: path.file_stem().unwrap().to_str().unwrap().into(),
-            ret: i64::from_str_radix(&ret, 10).unwrap(),
+            ret: isize::from_str_radix(&ret, 10).unwrap(),
             code: data,
         }
     }
 
-    pub fn run(&self, dump: bool) -> ExecutionResult<i64>
+    pub fn run(&self, dump: bool) -> ExecutionResult<isize>
     {
         let mut bc_mod = match generate_byte_code(&self.code, dump)
         {
@@ -51,7 +51,7 @@ impl Test
 }
 
 
-fn run_test(prog: &Path, dump: bool) -> ExecutionResult<i64>
+fn run_test(prog: &Path, dump: bool) -> ExecutionResult<isize>
 {
     let test = Test::load(prog);
     let ret = test.run(dump)?;
